add and remove occupation finished

+++ src/main/java/bean/OccupationBean.java
@@ -10,7 +10,9 @@ import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 import javax.faces.view.ViewScoped;
 import javax.inject.Named;
+import java.beans.BeanProperty;
 import java.io.Serializable;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -30,13 +32,10 @@ public class OccupationBean implements Serializable {
     public void init() {
         this.occupation = new Occupation();
         this.occupationDao = new OccupationDao();
+        this.selectedOccupationList = Arrays.asList();
         this.activeOccupationList = this.occupationDao.list().stream()
                 .filter(o -> o.getActiveStatus()).
                 collect(Collectors.toList());
-    }
-
-    public void newOccupation() {
-        this.occupation = new Occupation();
     }
 
     public void saveOccupation() {
@@ -51,16 +50,20 @@ public class OccupationBean implements Serializable {
         PrimeFaces.current().ajax().update("form:messages");/*add table update*/
     }
 
-    public String deleteButtonMessage() {
-        if (this.selectedOccupationList.isEmpty())
+    public String getDeleteButtonMessage() {
+        if (this.getHasOccupationsSelected())
             return this.selectedOccupationList.size() + " occupations selected";
         else
             return "delete";
     }
 
     public void hardRemove() {
-        this.selectedOccupationList.forEach(o -> this.occupationDao.remove(o.getId()));
+        if(this.getHasOccupationsSelected())
+            this.selectedOccupationList.forEach(o -> this.occupationDao.remove(o.getId()));
+    }
 
+    public Boolean getHasOccupationsSelected() {
+        return !this.selectedOccupationList.isEmpty();
     }
 
     public void setInactive() {
